Renames routes module to metrics
Fix #216
