+++ metrics/tempo_config.py
@@ -1,7 +1,7 @@
 """ Config module"""
 import pandas as pd
 
-from routes.date_utils import lookBack
+from metrics.date_utils import lookBack
 
 START_DATE = pd.Timestamp("2021-01-01")
 TODAY = pd.Timestamp("today")
